Create Updated Code
This is for the changes we are going to do to make the code our own! Please only save/commit to this document if you make significant changes and ONLY write comments about new things that differ from the original code. If you want to put comments about how the original code works, please use the OTHER file